+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Allowed headings:
 
 ## [Unreleased]
 
+### Additions
+
+* Added new `MiniAdvertCardView` component.
+
 ## [4.4.1] - 2023-02-03
 
 ### Changed